# src/Account-Api/Account.Api/Documentation/Swagger/Definition/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Account API",
    "description": "API that provides endpoints for account services.",
    "contact": {
      "name": "Filipi Mosquini",
      "email": "mosquinilabs@gmail.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "v1"
  },
  "paths": {
    "/api/accounts": {
      "post": {
        "tags": [
          "Account"
        ],
        "summary": "api/accounts. (Auth)",
        "description": "<p>\r\n  <b> Description: </b>\r\n  \r\n\r\n            This method create an account. \r\n</p>\r\n<p>\r\n  <b> Requirements: </b>\r\n  \r\n\r\n            Is required to be authenticated. \r\n</p>",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccountRequest"
              },
              "example": {"holder":"José da Silva"}
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccountRequest"
              },
              "example": {"holder":"José da Silva"}
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccountRequest"
              },
              "example": {"holder":"José da Silva"}
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAccountDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAccountDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAccountDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request\r\n           <ul><li>Document.IsInvalid</li><li>Document.IsRequired</li><li>Document.MustBeInformed</li></ul>"
          },
          "401": {
            "description": "Unauthorized\r\n           <ul><li>Blocked.UserUnauthorized</li></ul>"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Conflict\r\n           <ul><li>Idempotency.InProgress</li><li>Idempotency.KeyReuse</li></ul>"
          },
          "500": {
            "description": "InternalServerError\r\n           <ul><li>Error.Unexpected</li></ul>"
          },
          "503": {
            "description": "ServiceUnavailable\r\n           <ul><li>Idempotency.Unavailable</li></ul>"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/accounts/deactivate": {
      "patch": {
        "tags": [
          "Account"
        ],
        "summary": "api/accounts/deactivate. (Auth)",
        "description": "<p>\r\n  <b> Description: </b>\r\n  \r\n\r\n            This method deactivate an account. \r\n</p>\r\n<p>\r\n  <b> Requirements: </b>\r\n  \r\n\r\n            Is required to be authenticated. \r\n</p>",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found\r\n           <ul><li>Account.NotFound</li></ul>"
          },
          "500": {
            "description": "InternalServerError\r\n           <ul><li>Error.Unexpected</li></ul>"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/movements": {
      "post": {
        "tags": [
          "Movement"
        ],
        "summary": "api/movements. (Auth)",
        "description": "<p>\r\n  <b> Description: </b>\r\n  \r\n\r\n            This method create a movement. \r\n</p>\r\n<p>\r\n  <b> Requirements: </b>\r\n  \r\n\r\n            Is required to be authenticated. \r\n</p>",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMovementRequest"
              },
              "example": {"accountNumber":123,"amount":300,"movementType":67}
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMovementRequest"
              },
              "example": {"accountNumber":123,"amount":300,"movementType":67}
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMovementRequest"
              },
              "example": {"accountNumber":123,"amount":300,"movementType":67}
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request\r\n            <ul><li>Inactive.Account</li><li>Invalid.AccountNumber</li><li>Invalid.Amount</li><li>Invalid.MovementType</li><li>Invalid.RequestId</li></ul>"
          },
          "401": {
            "description": "Unauthorized\r\n           <ul><li>Blocked.UserUnauthorized</li></ul>"
          },
          "404": {
            "description": "Not Found\r\n            <ul><li>NotFound.Account</li></ul>"
          },
          "409": {
            "description": "Conflict\r\n           <ul><li>Conflict.OnlyCreditMovementAccepted</li><li>Idempotency.InProgress</li><li>Idempotency.KeyReuse</li></ul>"
          },
          "500": {
            "description": "InternalServerError\r\n            <ul><li>Error.Unexpected</li></ul>"
          },
          "503": {
            "description": "ServiceUnavailable\r\n           <ul><li>Idempotency.Unavailable</li></ul>"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/movements/balances": {
      "get": {
        "tags": [
          "Movement"
        ],
        "summary": "api/movements/balances. (Auth)",
        "description": "<p>\r\n  <b> Description: </b>\r\n  \r\n\r\n            This method get a balance result. \r\n</p>\r\n<p>\r\n  <b> Requirements: </b>\r\n  \r\n\r\n            Is required to be authenticated. \r\n</p>",
        "parameters": [
          {
            "name": "AccountNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetBalanceQueryDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetBalanceQueryDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetBalanceQueryDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request\r\n           <ul><li>Inactive.Account</li><li>Invalid.AccountNumber</li></ul>"
          },
          "401": {
            "description": "Unauthorized\r\n           <ul><li>Blocked.UserUnauthorized</li></ul>"
          },
          "404": {
            "description": "Not Found\r\n            <ul><li>NotFound.Account</li></ul>"
          },
          "500": {
            "description": "InternalServerError\r\n           <ul><li> Error.Unexpected </li></ul>"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "CreateAccountDto": {
        "type": "object",
        "properties": {
          "accountNumber": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "CreateAccountRequest": {
        "type": "object",
        "properties": {
          "holder": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateMovementRequest": {
        "type": "object",
        "properties": {
          "accountNumber": {
            "type": "integer",
            "description": "The bank account number",
            "format": "int32",
            "example": 123
          },
          "amount": {
            "type": "number",
            "description": "The amount of bank movement",
            "format": "double",
            "example": 300
          },
          "movementType": {
            "$ref": "#/components/schemas/MovementTypeEnum"
          }
        },
        "additionalProperties": false
      },
      "GetBalanceQueryDto": {
        "type": "object",
        "properties": {
          "accountNumber": {
            "type": "integer",
            "format": "int32"
          },
          "holder": {
            "type": "string",
            "nullable": true
          },
          "queryDate": {
            "type": "string",
            "format": "date-time"
          },
          "balance": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "MovementTypeEnum": {
        "enum": [
          67,
          68
        ],
        "type": "integer",
        "format": "int32"
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Standard Authorization header using the Bearer scheme (JWT). Example: \"bearer {token}\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}